Partial polyfill for CSS 'visibility'
Implements a partial polyfill for non-inheritable 'visibility'. This
does not polyfill the web behavior of being able to selectively override
the 'visibility' of elements nested within an element with
'visibility:hidden'.

As result, this patch also removes 'visibility' from the inheritable
text styles. Originally visibility was added here simply because it is
considered an inheritable property on web, even though we didn't do
anything with the property on native. In the context of these polyfills,
we're only interested in being able to inherit text styles and cannot do
anything with inherited visibility, so there is no functional change to
the existing polyfill for inheritable text styles.

Also note that no attempt is made to implement 'collapse' as per W3C
spec, as this behavior does not appear to be supported for flex items in
Chrome anyway (which behave the same whether 'collapse' or 'hidden' is
used.)