Share the same Animated.Value amongst transitions that happen in the same commit
Right now, updating state which affects transitioning properties accross
multiple components results in each component creating its own
Animated.Value and starting the animation in an effect.

This results in animations that are not in sync as they should be.

In this PR, transitions with the same delay, duration, timing function share
the same Animated.Value as long as they're part of the same commit.

useLayoutEffect is used to either create the Animated.Value or find an
existing suitable one to use.

Since all layout effects run before effects, we can assume that by the
time any component's effect runs, all Animated.Values have been created
or shared.

One of the component's effects will then kick off the animation and
empty the shared animation config map so that future commits don't
reuse any of the Animated.Values that were created. The other
component's effects will be no-ops.

Finally, reference counting is used to ensure only the last component
that unmounts would stop the animation if it was still running since
now they are shared it would be disruptive to stop the animation
unless the component was the last one relying on it