Add native support for CSS variables with fixed names
StyleX on web has a feature that prevents auto-generation of scoped CSS
var names created by `defineVars`. If the user-provided var key is a
string that starts with `--`, the key is preserved as-is and its value
can even be redefined in other `defineVars` calls.

Adding support for this on native involves removing the legacy feature
that converted kebab-case variable names to camelCase. Therefore, it's a
breaking change that requires users to modify to the `customProperties`
map (adding `--` prefix to all keys) passed to any use of the legacy
`ThemeProvider` context.

Close #73
Fix #50