Fix scaling of numeric fontSize
Previously fontSize was not scaled if the value is a number. The
viewport scaling was performed for CSS lengths but not specifically for
the fontSize or lineHeight property. Numeric lineHeight should not exist
as we convert it to a string in the preprocessing step. And numeric
fontSize is complicated by the multi-pass processing of styles which
means fontSize could be incorrectly scaled multiple times.

This patch only reprocesses non-numeric fontSize values, and introduces
a new resolving option that is used to pass around numeric fontSize (for
use in resolving logic) without treating it as a style that needs
further processing.