Go back to using ref callback instead of useImperativeHandle
We shouldn't create an object if the underlying node is null, as this
handle is meant to emulate the node but can't proxy methods if none
exist. Since useImperativeHandle isn't reset when the underlying node
changes, we go back to using ref callbacks. Rather than mutating the
underling node, we create a new object that wraps the relevant fields.