Add error messages for styles that only work on flex children
These messages will help developers on native to know when styles will
have no effect on web because they only work on flex children, but the
parent is not a flex container. On native, the styles will work because
all elements are `display:flex`, but this is not the case on web.