[wip] add react-strict-animated library
Note: still working on the build/package structure so not quite ready for review yet

This introduces a new package to the RSD repo that provides a subset of React Native's API that works on both web and native, and built with integration of react-strict-dom in mind. The native implementation is largely a thin passthrough to the proper Animated API and some settings pre-configured (such as `useAnimatedDriver` always being enabled).

The web side is a new Animated implementation that ends up being driven by the Web Animations API.